color change

+++ src/model/s21_aff_transform.cc
@@ -9,7 +9,7 @@ void AffTransform::clear() {
 
 int AffTransform::Scale(double koff) {
   if (koff != 0.0) {
-    for (int i = 0; i < obj_t_->count_of_vertexes * 3; i++) {
+    for (size_t i = 0; i < obj_t_->count_of_vertexes * 3; i++) {
       obj_t_->vertex_vector[i] *= koff;
     }
   } else {
@@ -20,7 +20,7 @@ int AffTransform::Scale(double koff) {
 
 int AffTransform::ScaleX(double koff) {
   if (koff != 0.0) {
-    for (int i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
       obj_t_->vertex_vector[i] *= koff;
     }
   } else {
@@ -31,7 +31,7 @@ int AffTransform::ScaleX(double koff) {
 
 int AffTransform::ScaleY(double koff) {
   if (koff != 0.0) {
-    for (int i = 1; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 1; i < obj_t_->count_of_vertexes * 3; i += 3) {
       obj_t_->vertex_vector[i] *= koff;
     }
   } else {
@@ -42,7 +42,7 @@ int AffTransform::ScaleY(double koff) {
 
 int AffTransform::ScaleZ(double koff) {
   if (koff != 0.0) {
-    for (int i = 2; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 2; i < obj_t_->count_of_vertexes * 3; i += 3) {
       obj_t_->vertex_vector[i] *= koff;
     }
   } else {
@@ -53,7 +53,7 @@ int AffTransform::ScaleZ(double koff) {
 
 void AffTransform::MoveX(double mv) {
   if (mv != 0.0) {
-    for (int i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
       obj_t_->vertex_vector[i] += mv;
     }
   }
@@ -62,7 +62,7 @@ void AffTransform::MoveX(double mv) {
 
 void AffTransform::MoveY(double mv) {
   if (mv != 0.0) {
-    for (int i = 1; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 1; i < obj_t_->count_of_vertexes * 3; i += 3) {
       obj_t_->vertex_vector[i] += mv;
     }
   }
@@ -70,14 +70,14 @@ void AffTransform::MoveY(double mv) {
 
 void AffTransform::MoveZ(double mv) {
   if (mv != 0.0) {
-    for (int i = 2; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 2; i < obj_t_->count_of_vertexes * 3; i += 3) {
       obj_t_->vertex_vector[i] += mv;
     }
   }
 }
 
 // void shift_dot(ObjT* obj_t_, dot_t dot) {
-//   for (int i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
+//   for (size_t i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
 //     obj_t_->vertex_vector[i] += dot.delta_x;
 //     obj_t_->vertex_vector[i + 1] += dot.delta_y;
 //     obj_t_->vertex_vector[i + 2] += dot.delta_z;
@@ -87,7 +87,7 @@ void AffTransform::MoveZ(double mv) {
 void AffTransform::TurnX(double angle) {
   if (angle != 0.0) {
     angle *= RAD_TO_GRAD;
-    for (int i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
       temp_y_ = obj_t_->vertex_vector[i + 1];
       temp_z_ = obj_t_->vertex_vector[i + 2];
       obj_t_->vertex_vector[i + 1] =
@@ -101,7 +101,7 @@ void AffTransform::TurnX(double angle) {
 void AffTransform::TurnY(double angle) {
   if (angle != 0.0) {
     angle *= -RAD_TO_GRAD;
-    for (int i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
       temp_x_ = obj_t_->vertex_vector[i];
       temp_z_ = obj_t_->vertex_vector[i + 2];
       obj_t_->vertex_vector[i] = temp_x_ * cos(angle) + temp_z_ * sin(angle);
@@ -114,7 +114,7 @@ void AffTransform::TurnY(double angle) {
 void AffTransform::TurnZ(double angle) {
   if (angle != 0.0) {
     angle *= RAD_TO_GRAD;
-    for (int i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
+    for (size_t i = 0; i < obj_t_->count_of_vertexes * 3; i += 3) {
       temp_x_ = obj_t_->vertex_vector[i];
       temp_y_ = obj_t_->vertex_vector[i + 1];
       obj_t_->vertex_vector[i] = temp_x_ * cos(angle) + temp_y_ * sin(angle);
